+++ spec/chef/gitlab-ee/recipes/consul_spec.rb
@@ -139,7 +139,7 @@ describe 'consul' do
   end
 
   describe 'consul::service_postgresql' do
-    let(:chef_run) { ChefSpec::SoloRunner.new(step_into: %w(runit_service account)).converge('gitlab-ee::default') }
+    let(:chef_run) { ChefSpec::SoloRunner.new(step_into: %w(runit_service)).converge('gitlab-ee::default') }
 
     before do
       allow(Gitlab).to receive(:[]).and_call_original