+++ spec/fixtures/cookbooks/test_crond_job/metadata.rb
@@ -1,2 +1,0 @@
-name 'test_crond_job'
-depends 'crond'